added volume functionality

+++ QNL/app/src/main/java/net/imanolgomez/qnl/BasicElement.java
@@ -18,7 +18,7 @@ public class BasicElement {
     // Instance variables
     private  int mId;
     private  String mName;
-    private  double mVersion;
+    private  float mVersion;
 
     /**
      * @param id The RouteElement's request ID
@@ -28,7 +28,7 @@ public class BasicElement {
     public BasicElement(
             int id,
             String name,
-            double version) {
+            float version) {
 
         // Set the instance fields from the constructor
         this.mId = id;
@@ -44,11 +44,11 @@ public class BasicElement {
         return mName;
     }
 
-    public double getVersion() {
+    public float getVersion() {
         return mVersion;
     }
 
-    public void setVersion(double version){
+    public void setVersion(float version){
         this.mVersion = version;
     }
